Restructuring State and unit tests for the same class
The States include more functionality. They will work as a type of linked list, although there are multiple possible arrows for multiple symbols. NFA class has less to do with tracking states, but a builder class will have to build the correct structure and pass the beginning state to the NFA.