Finished unit tests for State, began writing JavaDoc
Added a few more unit tests for State to cover different possibilities. All methods should now be tested at least in some manner, so I will not be focusing on producing more, at least for the moment.

Included the Maven JavaDoc module to the pom file, so that JavaDoc html-files are created when the project is processed.

Wrote JavaDoc for classes State and NFA. Created stubs of JavaDoc for other classes and methods for later completion.