NFAGenerator supports parentheses + created PatternProcessor class
NFAGenerator now supports all the mandatory operations: *, &, | and ().
To tackle useful shorthands I began work on PatternProcessor class. Its duty is to replace useful notation so that the pattern only contains the aforementioned basic operation symbols. The resulting string might be significantly longer than the input, but it will be simple in terms of present operation symbols. PatternProcessor supports operations +, * and [min, max] at the moment, but NFAGenerator doesn't use the class yet. Code is messy and tests nonexistent, but first priority is to get the functionality in place asap.