Added support for '-' shorhand in PatternProcessor. Also minor fixes + changes and tests
PatternProcessor also supports '-' operation: easy way to indicate a set of values, any of which will do. There were some bugs in the PatternProcessor class: for instance, in the case of repetitions if one limit was missing, the returned result was wrong.
Changed the negation symbol from ^ to ! in code. Reasons: easier to write (no need to press space), similar to the Java operation.
Created a great many tests for the PatternProcessor class. Added preemptively tests for NFACreator class when negation is present in the pattern string. Created stubs for testing for different shorthands.
Took first preparations towards processing negation: created DFAGenerator class, which will be responsible for turning an NFA into a DFA and then swappings its accepting and rejecting states. I will probably not implement DFAs as their own class, since DFAs are NFAs; I was thinking of using an attribute in NFA class to indicate whether the object is DFA. This information will be useful in turning NFA into DFA - if it is already DFA, there is no need for the extremely expensive operation of powerset construction.