Fixed a few issues relating to empty strings / empty symbols.
PatternProcessor handles empty strings: if the pattern to be processed is empty, elongateRegularExpression method replaces it with "#".

NFAGenerator skipped # characters entirely earlier. Now they are used to create the corresponding NFA and it's used to construct the whole NFA. All tests now pass except for the ones that test complement operations, which remain unimplemented