DFAGenerator's generateComplementDFA method works correctly
When given an NFA as input parameter, the method returns a DFA that recognizes exactly the strings that the parameter NFA does not. The method needs alphabet as a parameter, too.
The method doesn't use subset construction algorithm in straightforward way; all the states of the DFA are not created at the beginning (creating all would result in 2^n states), but new states are only created when a unique combination of the NFA's states are possible at some moment. If, for instance, there is no possibility of being in all the states of the NFA at once, the state corresponding to all the states won't be created in the DFA. This does not lower worst-case performance, but in many cases might save time and space.

Also created test class for the DFACreator. There aren't many tests yet, and some more complex cases will be tested when I've integrated DFACreator's functionality and handling the complement operator to the NFACreator.