With added support for complement operation, NFAGenerator is functionally completed
Although not particularly lightweight performancewise, the planned support for negation has been implemented. All the operations that I planned to support are now implemented. Also added plenty of tests for NFAGenerator. There could probably be more test cases for complex patterns that combine most of the operations, but to the extent that current tests offer the NFAGenerator works with any string that conforms to my idea of well structured regex pattern.