Wrote JavaDoc, refactored some code, created a stub for TextUI
Wrote JavaDoc for everything but the tests and the TextUI class. Divided some long methods into smaller components. Changed the method for adding states that are reachable by empty transitions - now same states are not considered more than twice.
Created a simple loop asking for input in the TextUI class.