Implemented own stack data structure
Created classes OwnStack and Node; the latter is needed as part of stack. The stack works as a singly linked list, so it contains a pointer to the topmost node. Each node in turn contains some data to be stored and a pointer to the node below them; the one at the bottom has a pointer to null. Basic operations of pop, push, and peek are supported.

Also created unit tests for both classes.

Have not replaced stacks in use with this implementation.