Sinificantly improved generating complement automata from DFA
By adding an attribute to the NFA class it's possible to indicate all the NON-accepting states with the accepted states set attribute. If DFAGenerators complement method receives an NFA that is certain to be DFA (it has been generated by that method at some point), the method simply changes the value of the inverted bit in it and returns the very same NFA. The NFA now accepts exactly the opposite language than it did before, but the only required change was that one bit.

The solution doesn't improve the general performance much, but at least some tricky inputs take less time.