Created own Hash Map class and tests
Created a Hash Map class that handles collisions with doubly linked collision lists, for which I created another class. Also created a third class, this one representing nodes in the collision lists: they contain key, value and two pointers to other nodes.

Wrote some tests for each of the three classes. The hash map is not finished, but the main functionality is present: putting, rehashing when needed, getting and removing. Methods AddAll and containsKey have to be implemented before the default Java class can be replaced.