Implemented OwnSet and replaced all imported data structures
Changed the implementation of several methods in HashSet, which now inherits most of its behavior and structure from HashTable class. HashSet also inherits the same class, since it's basically a Hash Map that only cares about the key values. Replaced all uses of the imported default implementations of both HashMap and HashSet with my own implementations. The fix for preventing multiple values for one key is now done in HashTable, which is a quick fix and should probably be moved inside Linked List.

Documentation and tests are lacking, since I was first concerned with just getting the tests to pass with the replaced data structures.