More extensive tests and documentation
Wrote tests for HashSet, OwnHashMap and SetIterator. Added JavaDoc everywhere it was lacking. Also changed the location of the apidocs from regex_parser directory to documentation directory.
Added pit test results to the documentation directory, and a link to the results to the testing document. Began work on week report 5.