Added first performance indicators
Text UI now offers a choice between simple comparisons between my regex parser implementation and that of Java default one. Also, when matching patterns with test strings, the milliseconds that it takes to generate a corresponding NFA and running the test string on it are displayed.

Fixed a potential path to IndexOutOfBoundsException in HashTable; clear method did not update capacity, so it was possible to generate too large indexes with scaledHashCode method if a large hash table had been cleared.