Caching of NFA simulations
When a test string is processed in an NFA, all the transitions are stored for later use. If the automaton has a set of current states that it has been in at some point, and reads a symbol that it has read in that configuration, it need not calculate the next states; it can simply look it up from the cache. Quick testing indicates that performance improves significantly is the same NFA is used to process the same input string, and probably improves even in cases when the same NFA is used for different test strings.

Fixed an oversight in the equals method in OwnSet, which previously led to null pointer exception at times. Also added copy methods for OwnSet and OwnHashMap, since the caching in NFA is vulnerable if all stored values are not copies, the only references to which are in that cache.