Began implementing the escape character
Modified the pattern processor so that leaves the escape characters and the characters following them alone, or at most wraps them in parentheses. NFAGenerator needs to be modified as well, but that remains entirely undone. NFA itself shouldn't need tweaking, but the alphabet has to be extended to account for more characters, e.g. +, [, ], ?.