Full support for escape characters
NFAGenerator will push the symbol following the escape character into the operand stack, so it will become a part of the generated NFA, and not be interpreted as operations. DFAGenerator's alphabet was extended, which will unfortunately increase space and time requirements for all negations. Wrote tests to examine how escape characters are handled.