Restructured packages, wrote documentation
Ran performance tests and added the results to the testing document. Began work on implementation document. Also altered state and all the methods and structures that work with empty transitions: both empty transitions and transitions that allow for any single character are stores as their own attributes in State. Added a search function to the text UI and probably plenty of minor tweaks.