UI and behavior changes
Changed program behavior when matching user selected regular expression against user selected test strings. Now the user will be prompted to give new test string for the same regular expression until they enter empty string twice in a row. After that they may exit the program or give another regular expression.

Gave the option to read any text file in the resources folder, not only Frankenstein. The user is prompted for a filename (including the file type), and empty input defaults to frankenstein.txt. The user is responsible for storing the file in the appropriate place, as well as ensuring that it contains only symbols that the regex parser supports.