Improved build process of project
Added plugins for creating a useful jar file and generating JavaDoc that reflects the state of the program when maven builds the project. In the .jar file are included the files from the resources folders, so the program should be runnable from it alone. There were significant troubles along the way, but everything seems to work in an identical fashion between running the .jar in terminal and running the main class in NetBeans...