Wrote documentation, minor fixes
OwnSet now offers any() method, which returns the first element it finds. This is used instead of iterator in DFAGenerator, which should improve performance - my iterator implementation goes through all the elements in a set before providing access to any of them. The work should be roughly in a finished state, although there remain some odd solutions and space for tuning performance. Also the program will probably not work with multiple escape characters in a row.